Migrate proof out
